Use int for trials; adjust chunk length
Change Trials field in MVEPGameConfiguration from float to int and expose an int property to represent discrete trial counts. Update default chunkLength in MVEPTimingConfiguration from 30f to 32f to tweak timing/chunk granularity. Note: callers that previously expected a float for Trials may need to be updated or cast accordingly.

+++ Neurogame/Assets/Scripts/MVEP/ScriptableObjects/MVEPGameConfiguration.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "MVEPGameConfig", menuName = "MVEP/MVEPGameConfig", order = 1)]
 public class MVEPGameConfiguration : ScriptableObject
 {
-  [SerializeField] private float trials;
-  public float Trials => trials;
+  [SerializeField] private int trials;
+  public int Trials => trials;
 
   [SerializeField] private float scoreMax;
   public float ScoreMax => scoreMax;
